+++ lookup.h
@@ -21,7 +21,7 @@ public:
 
   Square() : Square(0.5) {};
   Square(float duty);
-  int16_t next(int64_t tick, int32_t phi) override;
+  int16_t next(int32_t tick, int32_t phi) override;
 
 private:
 
@@ -34,7 +34,7 @@ class QuarterWtable : public Wavetable {
 
 public:
   
-  int16_t next(int64_t tick, int32_t phi) override;
+  int16_t next(int32_t tick, int32_t phi) override;
   
 protected:
 
@@ -69,7 +69,7 @@ class Triangle : public Wavetable {
 public:
 
   Triangle() = default;
-  int16_t next(int64_t tick, int32_t phi) override;
+  int16_t next(int32_t tick, int32_t phi) override;
 
 };
   
@@ -78,7 +78,7 @@ class HalfWtable : public Wavetable {
 
 public:
   
-  int16_t next(int64_t tick, int32_t phi) override;
+  int16_t next(int32_t tick, int32_t phi) override;
   
 protected:
 
@@ -105,7 +105,7 @@ class Saw : public Wavetable {
 public:
 
   Saw(float offset);
-  int16_t next(int64_t tick, int32_t phi) override;
+  int16_t next(int32_t tick, int32_t phi) override;
 
 private:
 
@@ -120,7 +120,7 @@ class FullWtable : public Wavetable {
 
 public:
   
-  int16_t next(int64_t tick, int32_t phi) override;
+  int16_t next(int32_t tick, int32_t phi) override;
   
 protected:
